Flask_AtikApp/requirements.txt
Flask==2.3.2
tensorflow==2.18.0
Pillow==9.5.0
numpy==2.0.0